Make Matrix2D handle negative coordinates.

+++ rstris/src/figure_face.rs
@@ -14,7 +14,7 @@ impl FigureFace {
             blocks: Matrix2D::new_init(blocks),
         }
     }
-    pub fn new_empty(width: usize, height: usize) -> FigureFace {
+    pub fn new_empty(width: u32, height: u32) -> FigureFace {
         FigureFace {
             blocks: Matrix2D::new(width, height, Block::new_not_set()),
         }
@@ -22,11 +22,11 @@ impl FigureFace {
     //
     // Returns a list of row index which contains at least one block
     //
-    pub fn get_row_with_blocks(&self) -> Vec<usize> {
-        let mut rows_with_blocks: Vec<usize> = Vec::new();
+    pub fn get_row_with_blocks(&self) -> Vec<u32> {
+        let mut rows_with_blocks: Vec<u32> = Vec::new();
         for y in 0..self.blocks.height() {
             for x in 0..self.blocks.width() {
-                if self.blocks.get(x, y).is_set() {
+                if self.blocks.get(x as i32, y as i32).is_set() {
                     rows_with_blocks.push(y);
                     break;
                 }
@@ -36,10 +36,10 @@ impl FigureFace {
     }
     pub fn get_block_positions(&self) -> Vec<Position> {
         let mut positions: Vec<Position> = Vec::new();
-        for y in 0..self.blocks.height() {
-            for x in 0..self.blocks.width() {
+        for y in 0..self.blocks.height() as i32 {
+            for x in 0..self.blocks.width() as i32 {
                 if self.blocks.get(x, y).is_set() {
-                    positions.push(Position::new(x as i32, y as i32));
+                    positions.push(Position::new(x, y));
                 }
             }
         }
@@ -47,33 +47,33 @@ impl FigureFace {
     }
     pub fn get_lowest_block(&self) -> Position {
         let mut pos = Position::new(i32::min_value(), i32::min_value());
-        for y in 0..self.blocks.height() {
-            for x in 0..self.blocks.width() {
-                if self.blocks.get(x, y).is_set() && y as i32 > pos.get_y() {
-                    pos = Position::new(x as i32, y as i32);
+        for y in 0..self.blocks.height() as i32 {
+            for x in 0..self.blocks.width() as i32 {
+                if self.blocks.get(x, y).is_set() && y > pos.get_y() {
+                    pos = Position::new(x, y);
                 }
             }
         }
         return pos;
     }
-    pub fn get_width(&self) -> usize {
+    pub fn get_width(&self) -> u32 {
         self.blocks.width()
     }
-    pub fn get_height(&self) -> usize {
+    pub fn get_height(&self) -> u32 {
         self.blocks.height()
     }
-    pub fn get_block(&self, x: usize, y: usize) -> &Block {
+    pub fn get_block(&self, x: i32, y: i32) -> &Block {
         &self.blocks.get(x, y)
     }
-    pub fn set_block(&mut self, x: usize, y: usize, block: &Block) {
+    pub fn set_block(&mut self, x: i32, y: i32, block: &Block) {
         self.blocks.set(x, y, block.clone());
     }
     pub fn place(&self, pf: &mut Playfield, pos: &Position) {
-        for row in 0..self.blocks.height() {
-            let pos_y = pos.get_y() + row as i32;
-            for col in 0..self.blocks.width() {
+        for row in 0..self.blocks.height() as i32 {
+            let pos_y = pos.get_y() + row;
+            for col in 0..self.blocks.width() as i32 {
                 let b = self.get_block(col, row);
-                let block_pos = Position::new(pos.get_x() + col as i32, pos_y);
+                let block_pos = Position::new(pos.get_x() + col, pos_y);
                 if b.is_set() && pf.contains(&block_pos) {
                     pf.set_block_by_pos(&block_pos, b.clone());
                 }
@@ -81,11 +81,11 @@ impl FigureFace {
         }
     }
     pub fn lock(&self, pf: &mut Playfield, pos: &Position) {
-        for row in 0..self.blocks.height() {
-            let pos_y = pos.get_y() + row as i32;
-            for col in 0..self.blocks.width() {
+        for row in 0..self.blocks.height() as i32 {
+            let pos_y = pos.get_y() + row;
+            for col in 0..self.blocks.width() as i32 {
                 let b = self.get_block(col, row);
-                let block_pos = Position::new(pos.get_x() + col as i32, pos_y);
+                let block_pos = Position::new(pos.get_x() + col, pos_y);
                 if b.is_set() && pf.contains(&block_pos) {
                     let mut b = b.clone();
                     b.state = BlockState::Locked;
@@ -95,11 +95,11 @@ impl FigureFace {
         }
     }
     pub fn remove(&self, pf: &mut Playfield, pos: &Position) {
-        for row in 0..self.blocks.height() {
-            let pos_y = pos.get_y() + row as i32;
-            for col in 0..self.blocks.width() {
+        for row in 0..self.blocks.height() as i32 {
+            let pos_y = pos.get_y() + row;
+            for col in 0..self.blocks.width() as i32 {
                 let b = self.get_block(col, row);
-                let block_pos = Position::new(pos.get_x() + col as i32, pos_y);
+                let block_pos = Position::new(pos.get_x() + col, pos_y);
                 if b.is_set() && pf.contains(&block_pos) {
                     pf.clear_block(&block_pos);
                 }
@@ -108,10 +108,10 @@ impl FigureFace {
     }
     pub fn test_collision(&self, pf: &Playfield, pos: &Position) -> BlockState {
         let mut state = BlockState::NotSet;
-        for row in 0..self.blocks.height() {
-            for col in 0..self.blocks.width() {
-                let block_pos = Position::new(pos.get_x() + col as i32,
-                                              pos.get_y() + row as i32);
+        for row in 0..self.blocks.height() as i32 {
+            for col in 0..self.blocks.width() as i32 {
+                let block_pos = Position::new(pos.get_x() + col,
+                                              pos.get_y() + row);
                 if self.get_block(col, row).is_set() {
                     let pf_block_state =
                         pf.get_block_by_pos(&block_pos).state.clone();
